+++ core/src/main/scala/codesearch/core/search/Search.scala
@@ -11,12 +11,12 @@ import codesearch.core.index.directory.СindexDirectory
 import codesearch.core.index.repository.Extensions
 import codesearch.core.search.Search.{CSearchPage, CSearchResult, CodeSnippet, Package, PackageResult, snippetConfig}
 import codesearch.core.search.SnippetsGrouper.SnippetInfo
-import codesearch.core.util.Helper
 import codesearch.core.util.Helper.readFileAsync
 import fs2.{Pipe, Stream}
 import io.chrisdavenport.log4cats.SelfAwareStructuredLogger
 import io.chrisdavenport.log4cats.slf4j.Slf4jLogger
 import codesearch.core.regex.space.SpaceInsensitive
+import codesearch.core.regex.RegexHelper
 
 import scala.sys.process.Process
 
@@ -85,7 +85,7 @@ trait Search {
     }
 
     val query: String = {
-      val preciseMatch: String = if (request.preciseMatch) Helper.preciseMatch(request.query) else request.query
+      val preciseMatch: String = if (request.preciseMatch) RegexHelper.preciseMatch(request.query) else request.query
       if (request.spaceInsensitive) SpaceInsensitive.spaceInsensitiveString(preciseMatch) else preciseMatch
     }
