+++ web-server/app/views/resultFrame.scala.html
@@ -1,3 +1,4 @@
+@import codesearch.core.regex.RegexHelper
 @import codesearch.core.util.Helper
 @import scala.collection.mutable
 @import scala.collection.mutable.ListBuffer
@@ -19,7 +20,7 @@
 )
 
 @regex = @{
-    Helper.buildRegex(query = query, space = space, insensitive = insensitive, precise = precise)
+    RegexHelper.buildRegex(query = query, space = space, insensitive = insensitive, precise = precise)
 }
 
 <section class="container">
